fix form submission bug
