+++ src/main/frontend/app/components/ParkShowContainer.js
@@ -2,6 +2,7 @@ import React, { useState, useEffect, Fragment } from "react"
 import ParkShow from "./ParkShow"
 import ReviewFormContainer from "./ReviewFormContainer"
 import ReviewShow from "./ReviewShow"
+import ParkReviewContainer from "./ParkReviewContainer"
 
 const ParkShowContainer = (props) => {
   const { state, id } = props.match.params
@@ -40,7 +41,7 @@ const ParkShowContainer = (props) => {
       </div>
       <div className="wrapper-individual-pet">
         <ParkShow key={park.id} data={park}/>
-        <ReviewFormContainer 
+        <ParkReviewContainer 
           park={park}
         />
       </div>
